modularize desktop shell actions hook
